Adding the now_next flags on talk lists for the API
The now_next flags need to be added regardless of whether the event is on,
fixed this in the talk model.  When events weren't on, the helper wasn't
returning the data as it should have.  Talk detail also needed the now_next
flags and these weren't being included so added these too